Include player steam_name getter.
Instead of storing names in the DB, we can track the users' name. Temporary fix for now for the users who have Unicode characters in their name (like the common ESEA star). Instead, we will store their name ASCII encoded, so we do not break the database? Or a different encoding that strips any bad characters.