Added in individual scoreboard call
This call will be used to return JSON values to update a scoreboard based on the stream. Also updated the match pages to start using FontAwesome instead of Glyphicon because it is no longer supported. Fixed to using regex for the sv_password, instead of relying on weird string fixes.