Leaderboard #21 (#28)
* Update match.html

Sorting on kills?

* Update match.html

Added sanity check as some servers are also likely to not have GoTV enabled on their server. Need to adjust match config as well to check if the port number is empty, or if tv_enabled is 0.

* Update match.html

* Update match.py

* Update match.html

Please Travis just spare me some mercy

* Update match.html

Have mercy 2.0 aka i hate myself.

* Update match.html

Attempting to sort again.

* Update match.html

Yay. Adjusting None values again.

* Update match.py

Adjusting gotv string again.

* Update README.md

* Adjusted sort order.

* Using temp workaround

For team logos. Now using TeamLogoManager plugin, due to an existing issue with Get5. However, in order to connect to the server you must have sv_allowupload 1 or you will not be able to connect.

* Super basic leaderboard based off of wins only.

Going to update with losses, and round difference as well.

* Updated to include round diff.

Math is wrong, and this is really inefficient.

* Adjust the team standings round diff.

Messed up and didn't keep adding the current score. This should fix it.

* Update leaderboard.py