Changes to leaderboard and matches.
Leaderboard will now link to the teams. Moved TeamLeaderboard class from leaderboard to models, left todo to create a better class for the leaderboard instead of using dictionaries. Updated util to use command to receive values from RCON. Updated match to enforce get5 auth, so that when a match is scheduled through the web panel it's forced to be authenticated.