Beginning start of demo uploads.
This includes modifying the api plugin to get a new forward which will send the filename to the server.

Also modified the apistats to account for the change of logo directories to follow the FastDL standard that is now in place.

Will have to revisit to look at SVG plugins.