Update Demo Uploads add Max Season Limits (#94)
* Added audit table for match rcons.

* Ensure autoincrement.

* Included one more audit trail.

* Included public team list.

* Included team audit table.

* Team Logo Uploading Complete.

You must be admin in order to upload a team logo, and it will automagically set once the users are saved. It tries to convert the image to a 64x64 image, and checks to see if under 10KB of size, if not, we error out. It's all or nothing.

* Uploading team logos now complete for admins.

* Added stats page to team layout.

* Logo Directory Updates

Not using a full path, I hope this doesn't break anything.

* Whoops

I'm a smart developer.

* Update __init__.py

* Update __init__.py

* Update __init__.py

* Update logo save logic.

* Append checks for file size again.

* Update to move logos to a fastDL standard.

* Basic database start for enc and logo fixes.

* inclued model update.

* Revert "inclued model update."

This reverts commit 0e2bae12b3012b54e9d1d8247257b601b35328d3.

* Revert "Basic database start for enc and logo fixes."

This reverts commit ce667d2f6b911db420628bf08ad968910206e91d.

* Revert WTForms version

This may have been causing some issues with forms not being able to auto-fill.

* Revert the reverts.

Turns out issue may have been Docker. Going back to where I was in hopes everything is okay.

* Updating logo issues that were found earlier.

Also updating db upgrade for enc on rcon password.

* Model changes and front end complete for SSH username password and boolean value.

Everything working so far, no forms issues.

* Include crypto resources in utils.

* Added in config values too

* RCON Password and SSH password are now encrypted by application!

* Migration script fix.

* Updating travis configs for test.

* fix var.

* Decrypt server pass on retrieval

* Update imports.

* Rcon password length increase

* Update imports.

* Fixed(?) testing libraries for server.

* Include encryption test.

* Unencrypt rcon when sending message to server.

* Server availability issue fixed?

* Oops.

* Changing decrypt to return.

* Beginning start of demo uploads.

This includes modifying the api plugin to get a new forward which will send the filename to the server.

Also modified the apistats to account for the change of logo directories to follow the FastDL standard that is now in place.

Will have to revisit to look at SVG plugins.

* Updated screenshots.

* Remove demo upload start. Need to do it via plugin.

* Include executable config that is now generated.

* Missed removal of SSH

* Missed

Again, removing encSSH

* Update README.md

* Attempting to create demorecording.

This will require a new file for loading FTP connections.

MUST REQUIRE cURL EXTENSION ON SERVER.

TO COMPILE YOU MUST HAVE TEASYFTP AND CURL. I will probably revert this due to it most likely not working.

* Remove use of tEasyFTP

Included use of System2. Now a system MUST require System2 extension if they wish to demo upload.

* Updated the demo upload test. Left msgs for console.

* Adjusted logo for the website to remove branding.

* Removing branding.

* Updated gitignore to remove branding.

* Create config value to enable/disable FTP upload.

* Adjusted if statement, awaiting reply to see why system2 fails to upload FTP.

* I spull good.

* Error check was failing. Print out any failure messages right now.

* Update README.md

* Updated plugin to support compression.

Updated fail case as data is more commonly misrepresented.

* Begin database implementation of preferred names.

* Dynamic team creation in the form has been completed.

* "Dynamically" retrieve data for forms. Super hacky and need to find a better fix.

In other news most of the backend/routing is complete for team size modifying. Next is the jinja template.

* Fix Match Stats Mobile View

It was going out of the mobile view section and going out of table area. Simple fix of adding scroll bar.

* Keep login consistent on pages.

* Finished inital testing of dynamic team limits.

* Updated stats to be consistent with panels.

* Database and saving preferred names complete.

Next is sending it off as JSON to the plugin.

* Finish preferred match name.

Start on API request for demos to store the filename.

* Adjust filter.

* Update time submit on forfeit.

* Send back proper filename and add to database.

Match page will now show a download button if demo exists.

Testing for demos needed.

* SVGs now able to upload to web panel. Will require testing.

Plugin still needs to be updated for this as well.

* Indentation fix.

* Adjust rendering.

* Update team.py

* Update logic on finding logos. Probably will fail.

* SVG Uploads now work.

May need to adjust viewable SVG, as it looks for width and height in root object.

* Adjusted folder locations and SVG upload now works

Adjusting demo uploads to allow for no compression.

Time to begin work on API stats logos.

* SVG Logos now completed in plugin.

* Renaming to get5 webpanel to be more generic.

* Webpanel name now in config.

Header now changes via what is placed in the config file.

* Missed config value.

* Update issue templates (#77)

* Development to Master (#40)

Leaderboards and seasons are now in place. Scoreboard API for matches need to be updated to view multiple maps in BO series, but other than that, seasons should work now as intended.

* Update issue templates

* Fixed leaderboard counting.

* Fix seasons not showing all matches.

* Remove logging. Logos now upload properly.

* Update seasons test as new routing is working.

* Update dependency.

Didn't realized pycrypto was broken and vulnerable.

* Update season link for users.

* Fix teams not being able to be created if not admin.

* Include knifing for sides feature.

Satisifes #80. Please test thoroughly to ensure everything is working as intended.

* Missed formatting for remote server.

* Adjust some debug statements.

* Return success.

Fixes demos not saving in DB?

* Hack fix inbound.

The way get5 handles the post-match config is it removes the API values before DemoFinish can be called. This means we have to store the values and remove them ourselves too.

Updated match page to show full URL.

TODO: Move the URL into map_stats instead.

Cleaned up scripts folder as some were unneeded.

* Update readme and images.

* Removing Hack Fix

Not actually needed, as the cancel match command (which was being used to test) behaves differently than just a series ending. It restores cvars and then stops demo, instead of stopping demo then restoring cvars.

* Demo uploads can now compress.

* Change where demo URLs are saved.

* Remove API Cvar from config.

Not needed. Breaks ready system.

* Add in hack fix as finishing doesn't work as intended.

* Veto first fix.

* Fixed demo uploads.

* Remove debugging, demo uploads finally complete.

* Include max season check.

Fix max matches actually doing something.